# sort: immediately compare whole lines if they parse as numbers