@sylvestre 2025-03-25
Could you please run hyperfine with all the commands at once ?
see
https://github.com/uutils/coreutils/blob/main/docs/src/performance.md

and what clang/gcc have to do here? :)