@sylvestre 2025-03-25
well done
how did you generate shuffled.txt ? 

could you please add your benchmark: https://github.com/uutils/coreutils/blob/main/src/uu/sort/BENCHMARKING.md