@MoSal 2025-03-25
> how did you generate shuffled.txt ?

A simple
```
shuf -i 1-1000000 -n 1000000 > /tmp/shuffled.txt
```

or

```
seq 1 1000000 | sort -R > /tmp/shuffled.txt`
```