@sylvestre 2025-03-26
@MoSal could you please add your benchmark: https://github.com/uutils/coreutils/blob/main/src/uu/sort/BENCHMARKING.md :)
(not the result but the way you generated the file + the hyperfine command)